Rename json -> record since it's not json anymore
